# Performance: Optimize LCP from 301ms to 262ms