--- a/src/lib/posthog-lazy.ts
+++ b/src/lib/posthog-lazy.ts
@@ -540,3 +540,11 @@ export function getRateLimiterStats(): {
     },
   };
 }
+
+/**
+ * Initialize PostHog (for manual initialization)
+ * Returns a promise that resolves when PostHog is loaded
+ */
+export async function initPostHog(): Promise<PostHogInstance | null> {
+  return loadPostHog();
+}