## Summary
Optimized Largest Contentful Paint (LCP) by deferring non-critical JavaScript initialization, achieving a 13% improvement (39ms faster).

## Performance Results

| Metric | Before | After | Improvement |
|--------|--------|-------|-------------|
| **LCP** | 301ms | 262ms | **39ms (13% faster)** ✅ |
| **TTFB** | 2-6ms | 3ms | Maintained excellence ✅ |
| **CLS** | 0.00 | 0.00 | Perfect stability ✅ |

## Implementation

### Phase 1: Preconnect Optimization (23ms saved)
- Removed 4 unused preconnects exceeding browser recommendation
- Kept only critical Supabase preconnect
- Converted GitHub/fonts to cheaper `dns-prefetch`
- Deferred Web Vitals initialization to `requestIdleCallback`

### Phase 2: Progressive Features Deferral (14ms saved)
- Moved progressive capture features from 500ms timeout to `requestIdleCallback`
- Features load during browser idle time instead of blocking render

### Phase 3: PostHog Interaction Loading (2ms saved)
- PostHog now loads on first user interaction (mousedown, scroll, touch)
- 3-second fallback for non-interactive sessions
- Web Vitals starts with Supabase only, adds PostHog after load

## Files Changed
- `index.html` - Optimized preconnect/dns-prefetch hints
- `src/App.tsx` - Implemented deferred initialization patterns
- `src/lib/posthog-lazy.ts` - Added `initPostHog()` export function

## Testing
Verified with Chrome DevTools Performance profiling at each phase:
- ✅ TypeScript compilation passes
- ✅ Vite build successful
- ✅ Web Vitals monitoring functional
- ✅ Progressive capture working
- ✅ PostHog analytics capturing
- ✅ No visual regressions

## Technical Notes
- Follows `/docs/performance/PERFORMANCE_BEST_PRACTICES.md`
- Uses `requestIdleCallback` with fallbacks for browser compatibility
- Maintains all functionality while improving performance
- No breaking changes to user experience

Closes #940

🤖 Generated with [Claude Code](https://claude.com/claude-code)